requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
